Support tilt integration without sprockets
This adds a tilt/opal.rb file that allows you to render opal templates
via Tilt without using sprockets. It changes lib/opal/sprockets/processor.rb
to load tilt/opal.rb and augment it with sprockets specific code.

This allows you to use opal with asset engines that don't use sprockets.
